Add a spooky surprise!
Tile is credit (Bloax). Grunt & gammafunk also helped!

+++ crawl-ref/source/files.cc
@@ -1147,6 +1147,26 @@ static bool _leave_level(dungeon_feature_type stair_taken,
     return popped;
 }
 
+/**
+ * Warn the player that two ghost files have been loaded into the current
+ * level, resulting in somewhere between two and twenty ghosts being present.
+ *
+ * Warnings may be more spooky than actually useful.
+ *
+ * @return  A message that will send shivers down players' spines, assuming
+ *          they aren't in wisp form!
+ */
+static const char* _double_ghost_spookmessage()
+{
+    static const char* spookmessages[] = {
+        "You are filled with an overwhelming sense of foreboding!",
+        "You feel a terrible frisson of fear!",
+        "You are flooded with an inexplicable sense of dread!",
+        "You feel that you have entered a very terrible place...",
+        "There is something very spooky about this place!"
+    };
+    return RANDOM_ELEMENT(spookmessages);
+}
 
 /**
  * Generate a new level.
@@ -1186,14 +1206,22 @@ static void _make_level(dungeon_feature_type stair_taken,
     _clear_env_map();
     builder(true, stair_type);
 
+    const bool is_halloween = today_is_halloween();
+
     if (!crawl_state.game_is_tutorial()
         && !crawl_state.game_is_zotdef()
         && !Options.seed
         && !player_in_branch(BRANCH_ABYSS)
         && (!player_in_branch(BRANCH_DUNGEON) || you.depth > 2)
-        && one_chance_in(3))
+        && one_chance_in(is_halloween ? 2 : 3))
     {
-        load_ghost(true);
+        if (is_halloween && coinflip())
+        {
+            mpr(_double_ghost_spookmessage());
+            load_ghost(true);
+        }
+        const bool delete_ghost = !is_halloween || one_chance_in(3);
+        load_ghost(true, delete_ghost);
     }
     env.turns_on_level = 0;
     // sanctuary
@@ -1742,9 +1770,10 @@ static string _find_ghost_file()
  * Attempt to load one or more ghosts into the level.
  *
  * @param creating_level    Whether a level is currently being generated.
+ * @param delete_file       Whether to delete the ghost file after loading it.
  * @return                  Whether ghosts were actually generated.
  */
-bool load_ghost(bool creating_level)
+bool load_ghost(bool creating_level, bool delete_file)
 {
     const bool wiz_cmd = (crawl_state.prev_cmd == CMD_WIZARD);
 
@@ -1805,8 +1834,11 @@ bool load_ghost(bool creating_level)
     }
     inf.close();
 
-    // Remove bones file - ghosts are hardly permanent.
-    unlink(ghost_filename.c_str());
+    if (delete_file)
+    {
+        // Remove bones file - ghosts are hardly permanent.
+        unlink(ghost_filename.c_str());
+    }
 
     if (!debug_check_ghosts())
     {
